+++ src/logics/AaveV3/SimpleAaveLogic.ts
@@ -1,9 +1,9 @@
 import { Big } from "big.js";
-import configuration from "../configuration";
-import Token from "../token";
-import { typechain } from "../types";
-import type { Prettify } from "../util/types";
-import { AbstractRoutingLogic } from "./AbstractRoutingLogic";
+import configuration from "../../configuration";
+import Token from "../../token";
+import { typechain } from "../../types";
+import type { Prettify } from "../../util/types";
+import { AbstractRoutingLogic } from "../AbstractRoutingLogic";
 import { BaseAaveLogic } from "./BaseAaveLogic";
 
 /**